move world loading code into `Swarm.Game.World.Load`
