+++ swarm.cabal
@@ -133,12 +133,13 @@ library
                       Swarm.Game.Terrain
                       Swarm.Game.Value
                       Swarm.Game.World
+                      Swarm.Game.World.Abstract
                       Swarm.Game.World.Compile
                       Swarm.Game.World.Coords
                       Swarm.Game.World.Eval
                       Swarm.Game.World.Gen
-                      Swarm.Game.World.Abstract
                       Swarm.Game.World.Interpret
+                      Swarm.Game.World.Load
                       Swarm.Game.World.Parse
                       Swarm.Game.World.Syntax
                       Swarm.Game.World.Typecheck
